Move classes InstanceMenu and InstanceMenuItem to own file
